refactor: adjust imports for Utils references
