+++ GitHubClients/GitClient.ts
@@ -5,7 +5,7 @@ import { PageInfoModel } from "../deps.ts";
 import { GitBranchModel } from "../deps.ts";
 import { RawRefsGetBranchModel } from "../deps.ts";
 import { RawGitBranchModel } from "../deps.ts";
-import { Utils } from "../core/Utils.ts";
+import { Utils } from "../deps.ts";
 import { RepoClient } from "./RepoClient.ts";
 import { getCreateBranchMutation } from "../core/GraphQl/Mutations/CreateBranchMutation.ts";
 import { addCommitMutation } from "../core/GraphQl/Mutations/AddCommitMutation.ts";
