+++ GitHubClients/IssueClient.ts
@@ -2,7 +2,7 @@ import { Guard } from "../core/Guard.ts";
 import { LabelClient } from "./LabelClient.ts";
 import { IssueModel } from "../deps.ts";
 import { LabelModel } from "../deps.ts";
-import { Utils } from "../core/Utils.ts";
+import { Utils } from "../deps.ts";
 import { GitHubHttpStatusCodes, IssueOrPRState } from "../core/Enums.ts";
 import { GitHubClient } from "../deps.ts";
 import { IssueOrPRRequestData } from "../core/IssueOrPRRequestData.ts";
