+++ GitHubClients/MilestoneClient.ts
@@ -2,7 +2,7 @@ import { Guard } from "../core/Guard.ts";
 import { IssueModel } from "../deps.ts";
 import { MilestoneModel } from "../deps.ts";
 import { PullRequestModel } from "../deps.ts";
-import { Utils } from "../core/Utils.ts";
+import { Utils } from "../deps.ts";
 import { GitHubHttpStatusCodes, IssueOrPRState, MergeState } from "../core/Enums.ts";
 import { GitHubClient } from "../deps.ts";
 import { IssueClient } from "./IssueClient.ts";
