+++ GitHubClients/OrgClient.ts
@@ -4,7 +4,7 @@ import { Guard } from "../core/Guard.ts";
 import { GitHubVarModel } from "../deps.ts";
 import { GitHubVariablesModel } from "../deps.ts";
 import { UserModel } from "../deps.ts";
-import { Utils } from "../core/Utils.ts";
+import { Utils } from "../deps.ts";
 import { OrganizationError } from "./Errors/OrganizationError.ts";
 
 /**
