+++ GitHubClients/ReleaseClient.ts
@@ -2,7 +2,7 @@ import { GitHubHttpStatusCodes } from "../core/Enums.ts";
 import { GitHubClient } from "../deps.ts";
 import { Guard } from "../core/Guard.ts";
 import { ReleaseModel } from "../deps.ts";
-import { Utils } from "../core/Utils.ts";
+import { Utils } from "../deps.ts";
 import { basename, existsSync } from "../deps.ts";
 import { ReleaseError } from "./Errors/ReleaseError.ts";
 import { ReleaseOptions } from "./ReleaseOptions.ts";
