+++ GitHubClients/TagClient.ts
@@ -1,6 +1,6 @@
 import { Guard } from "../core/Guard.ts";
 import { TagModel } from "../deps.ts";
-import { Utils } from "../core/Utils.ts";
+import { Utils } from "../deps.ts";
 import { GitHubHttpStatusCodes } from "../core/Enums.ts";
 import { GitHubClient } from "../deps.ts";
 import { TagError } from "./Errors/TagError.ts";
