+++ tests/UtilsTests.ts
@@ -1,5 +1,5 @@
 import { assertEquals } from "../deps.ts";
-import { Utils } from "../core/Utils.ts";
+import { Utils } from "../deps.ts";
 import { TestDataArgs } from "./TestTypes.ts";
 
 Deno.test("clamp |> when-num-is-larger-than-max |> returns-max-value", () => {
